Refactoring product overview page
